backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1
python-multipart==0.0.12
pandas==2.3.3
numpy==2.3.5
plotly==6.5.0
fpdf==1.7.2
openpyxl==3.1.5
python-dotenv==1.0.1
itsdangerous==2.2.0
